% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMLB.R
\name{getMLB}
\alias{getMLB}
\title{getMLB}
\usage{
getMLB()
}
\description{
This function calculates the most recent MLB playoff appearance date for all MLB teams based on data from Wikipedia \url{https://en.wikipedia.org/wiki/Major_League_Baseball}
}
\examples{
getMLB()
}
