% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNBA.R
\name{getNBA}
\alias{getNBA}
\title{getNBA}
\usage{
getNBA()
}
\description{
Calculates the most recent playoff appearance for every NBA team based on data from \url{https://www.basketball-reference.com}
}
\examples{
getNBA()
}
