% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNFL.R
\name{getNFL}
\alias{getNFL}
\title{getNFL}
\usage{
getNFL()
}
\description{
Calculates the most recent NFL playoff appearance date for all NFL teams based on data from Wikipedia \url{https://en.wikipedia.org/wiki/National_Football_League}
}
\examples{
getNFL()
}
