% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNHL.R
\name{getNHL}
\alias{getNHL}
\title{getNHL}
\usage{
getNHL()
}
\description{
Calculates most recent playoff appearence for all NHL teams from 1998 to 2018 from the site \url{https://www.hockey-reference.com}.
Does not pull data for 2004-05 due to the NHL strike
}
\examples{
getNHL()
}
