% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizePlayoffDrought.R
\name{summarizePlayoffDrought}
\alias{summarizePlayoffDrought}
\title{summarizePlayoffDrought}
\usage{
summarizePlayoffDrought(defunct = NULL)
}
\arguments{
\item{defunct}{Should defunct teams be omitted?}
}
\description{
Summarizes longest playoff drought by league for the 4 major US professional sports leagues
}
