% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizePlayoffDrought.R
\name{summarizePlayoffDrought}
\alias{summarizePlayoffDrought}
\title{summarizePlayoffDrought}
\usage{
summarizePlayoffDrought()
}
\description{
Summarizes longest playoff drought by league for the 4 major US professional sports leagues
}
